chore: move displaycommithash back to init
