+++ static/scripts/rewards/init.ts
@@ -2,8 +2,20 @@ import { app } from "./app-state";
 import { readClaimDataFromUrl } from "./render-transaction/read-claim-data-from-url";
 import { grid } from "./the-grid";
 
+displayCommitHash();
 grid(document.getElementById("grid") as HTMLElement, gridLoadedCallback); // @DEV: display grid background
 readClaimDataFromUrl(app).catch(console.error); // @DEV: read claim data from URL
+
+const footer = document.querySelector(".footer") as Element;
+footer.classList.add("animate");
+
+declare const commitHash: string; // @DEV: passed in at build time check build/esbuild-build.ts
+function displayCommitHash() {
+  // display commit hash in footer
+  const buildElement = document.querySelector(`#build a`) as HTMLAnchorElement;
+  buildElement.innerHTML = commitHash;
+  buildElement.href = `https://github.com/ubiquity/pay.ubq.fi/commit/${commitHash}`;
+}
 
 // cSpell:ignore llback
 function gridLoadedCallback() {
